Rename module: `reg` to `gpr`
